feat(size01): split templates.test.mjs into 3 partial test files [T001291]
